코드 2일차

+++ src/main/java/com/example/beenproject/common/ResVo.java
@@ -8,5 +8,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ResVo {
-    private Long result;
+    private int result;
 }
